# openapi/content-workflow.openapi.yaml
openapi: 3.1.0
info:
  title: Content Workflow API
  version: 0.1.0
  summary: MVP API for content ingestion, drafting, review, publishing, and analytics snapshots.
  description: |
    This specification covers the MVP workflow:
    ingestion -> normalization -> topic selection -> draft generation -> asset rendering -> review -> publishing -> metrics snapshots.

    Implementation constraints encoded in this spec:
    - `Draft.sourceRefs` is required and must contain at least one source reference before review submission.
    - `ChannelVariant` is a platform-specific content object and does not expose or reuse the draft body as the publish payload.
    - `POST /publish-tasks` and `POST /publish-tasks/{publishTaskId}/run` are idempotent and require `Idempotency-Key`.
    - `PerformanceMetric` records time-series snapshots with `capturedAt`; the API does not support in-place metric updates.
    - `RawItem.contentHash` and `NormalizedItem.dedupeKey` are both required to support raw-content and semantic deduplication separately.
servers:
  - url: https://api.example.com/api/v1
    description: Production
  - url: http://localhost:3000/api/v1
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /drafts:
    get:
      tags: [Drafts]
      summary: List drafts
      parameters:
        - in: query
          name: topicId
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/StatusQuery'
      responses:
        '200':
          description: Draft list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DraftListResponse'
  /drafts/generate:
    post:
      tags: [Drafts]
      summary: Generate a draft from a topic
      description: Generated drafts must include at least one source reference.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateDraftRequest'
      responses:
        '201':
          description: Draft created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Draft'
        '409':
          description: Draft cannot be generated because the topic lacks valid normalized inputs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /drafts/{draftId}:
    get:
      tags: [Drafts]
      summary: Get a draft
      parameters:
        - $ref: '#/components/parameters/DraftId'
      responses:
        '200':
          description: Draft
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Draft'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Drafts]
      summary: Update a draft
      parameters:
        - $ref: '#/components/parameters/DraftId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDraftRequest'
      responses:
        '200':
          description: Updated draft
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Draft'
  /drafts/{draftId}/regenerate:
    post:
      tags: [Drafts]
      summary: Regenerate a draft
      parameters:
        - $ref: '#/components/parameters/DraftId'
      responses:
        '202':
          description: Regeneration queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobAcceptedResponse'
  /drafts/{draftId}/submit-review:
    post:
      tags: [Drafts]
      summary: Submit a draft for review
      description: |
        Review submission fails unless the draft has at least one `sourceRefs` entry.
      parameters:
        - $ref: '#/components/parameters/DraftId'
      responses:
        '200':
          description: Draft moved to review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewTask'
        '409':
          description: Draft missing required source references
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /drafts/{draftId}/render:
    post:
      tags: [Assets]
      summary: Render assets for a draft
      parameters:
        - $ref: '#/components/parameters/DraftId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenderDraftRequest'
      responses:
        '202':
          description: Render queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobAcceptedResponse'
...
  /drafts/{draftId}/channel-variants:
    post:
      tags: [ChannelVariants]
      summary: Generate platform-specific variants for a draft
      description: |
        Channel variants are independent platform payloads.
        They intentionally exclude the draft body to prevent direct reuse of generic draft text as the publish payload.
      parameters:
        - $ref: '#/components/parameters/DraftId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChannelVariantsRequest'
      responses:
        '201':
          description: Variants created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelVariantListResponse'
  /channel-variants:
    get:
      tags: [ChannelVariants]
      summary: List channel variants
      parameters:
        - in: query
          name: draftId
          schema:
            type: string
            format: uuid
        - in: query
          name: platform
          schema:
            $ref: '#/components/schemas/PlatformType'
      responses:
        '200':
          description: Variant list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelVariantListResponse'
  /channel-variants/{channelVariantId}:
    get:
      tags: [ChannelVariants]
      summary: Get a channel variant
      parameters:
        - $ref: '#/components/parameters/ChannelVariantId'
      responses:
        '200':
          description: Channel variant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelVariant'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [ChannelVariants]
      summary: Update a channel variant
      parameters:
        - $ref: '#/components/parameters/ChannelVariantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateChannelVariantRequest'
      responses:
        '200':
          description: Updated channel variant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelVariant'
  /review-tasks:
    get:
      tags: [Review]
      summary: List review tasks
      parameters:
        - $ref: '#/components/parameters/StatusQuery'
        - in: query
          name: reviewerId
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Review task list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewTaskListResponse'
    post:
      tags: [Review]
      summary: Create a review task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewTaskRequest'
      responses:
        '201':
          description: Review task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewTask'
  /review-tasks/{reviewTaskId}:
    get:
      tags: [Review]
      summary: Get a review task
      parameters:
        - $ref: '#/components/parameters/ReviewTaskId'
      responses:
        '200':
          description: Review task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewTask'
        '404':
          $ref: '#/components/responses/NotFound'
  /review-tasks/{reviewTaskId}/approve:
    post:
      tags: [Review]
      summary: Approve a review task
      parameters:
        - $ref: '#/components/parameters/ReviewTaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApproveReviewTaskRequest'
      responses:
        '200':
          description: Review task approved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewTask'
  /review-tasks/{reviewTaskId}/reject:
    post:
      tags: [Review]
      summary: Reject a review task
      parameters:
        - $ref: '#/components/parameters/ReviewTaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RejectReviewTaskRequest'
      responses:
        '200':
          description: Review task rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewTask'
  /review-tasks/{reviewTaskId}/request-regenerate:
    post:
      tags: [Review]
      summary: Request regeneration from review
      parameters:
        - $ref: '#/components/parameters/ReviewTaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RejectReviewTaskRequest'
      responses:
        '200':
          description: Regeneration requested
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewTask'
  /publish-tasks:
    get:
      tags: [Publish]
      summary: List publish tasks
      parameters:
        - $ref: '#/components/parameters/StatusQuery'
        - in: query
          name: platform
          schema:
            $ref: '#/components/schemas/PlatformType'
      responses:
        '200':
          description: Publish task list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishTaskListResponse'
    post:
      tags: [Publish]
      summary: Create a publish task
      description: |
        This endpoint is idempotent. Clients must send `Idempotency-Key`.
        Replays with the same key and same effective payload return the original task.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePublishTaskRequest'
      responses:
        '201':
          description: Publish task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishTask'
        '200':
          description: Existing publish task returned for an idempotent replay
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishTask'
        '409':
          $ref: '#/components/responses/Conflict'
  /publish-tasks/{publishTaskId}:
    get:
      tags: [Publish]
      summary: Get a publish task
      parameters:
        - $ref: '#/components/parameters/PublishTaskId'
      responses:
        '200':
          description: Publish task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishTask'
        '404':
          $ref: '#/components/responses/NotFound'
  /publish-tasks/{publishTaskId}/run:
    post:
      tags: [Publish]
      summary: Run a publish task immediately
      description: |
        This endpoint is idempotent. Clients must send `Idempotency-Key`.
        Replays must not create duplicate outbound posts.
      parameters:
        - $ref: '#/components/parameters/PublishTaskId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202':
          description: Publish run queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishTask'
        '409':
          $ref: '#/components/responses/Conflict'
  /publish-tasks/{publishTaskId}/retry:
    post:
      tags: [Publish]
      summary: Retry a failed publish task
      parameters:
        - $ref: '#/components/parameters/PublishTaskId'
      responses:
        '202':
          description: Retry queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishTask'
  /publish-tasks/{publishTaskId}/cancel:
    post:
      tags: [Publish]
      summary: Cancel a queued publish task
      parameters:
        - $ref: '#/components/parameters/PublishTaskId'
      responses:
        '200':
          description: Publish task cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishTask'
...
  /dashboard/overview:
    get:
      tags: [Dashboard]
      summary: Get dashboard overview metrics
      parameters:
        - in: query
          name: date
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Dashboard overview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardOverview'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    DraftId:
      in: path
      name: draftId
      required: true
      schema:
        type: string
        format: uuid
...
    ChannelVariantId:
      in: path
      name: channelVariantId
      required: true
      schema:
        type: string
        format: uuid
    ReviewTaskId:
      in: path
      name: reviewTaskId
      required: true
      schema:
        type: string
        format: uuid
    PublishTaskId:
      in: path
      name: publishTaskId
      required: true
      schema:
        type: string
        format: uuid
    StatusQuery:
      in: query
      name: status
      schema:
        type: string
...
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 255
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Request conflicts with current resource state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    DraftType:
      type: string
      enum: [summary, deep]
    DraftStatus:
      type: string
      enum: [draft, rendered, in_review, approved, rejected, scheduled, published, failed]
    ReviewStatus:
      type: string
      enum: [pending, approved, rejected, regenerate_requested]
    PublishStatus:
      type: string
      enum: [queued, running, published, failed, cancelled]
    PlatformType:
      type: string
      enum: [xiaohongshu, douyin, tiktok, facebook]
...
    SourceReference:
      type: object
      additionalProperties: false
      required: [sourceId, normalizedItemId, title, url]
      properties:
        sourceId:
          type: string
          format: uuid
        normalizedItemId:
          type: string
          format: uuid
        title:
          type: string
        url:
          type: string
          format: uri
        publishedAt:
          type: string
          format: date-time
...
    Draft:
      type: object
      additionalProperties: false
      required:
        [id, topicId, draftType, languageCode, tone, outline, facts, sourceRefs, version, status, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        topicId:
          type: string
          format: uuid
        draftType:
          $ref: '#/components/schemas/DraftType'
        languageCode:
          type: string
        tone:
          type: string
        title:
          type: string
        outline:
          type: array
          items:
            type: object
            additionalProperties: true
        body:
          type: string
        facts:
          type: array
          items:
            type: object
            additionalProperties: true
        sourceRefs:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/SourceReference'
          description: Required before review submission.
        version:
          type: integer
          minimum: 1
        status:
          $ref: '#/components/schemas/DraftStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    GenerateDraftRequest:
      type: object
      additionalProperties: false
      required: [topicId, draftType, tone, languageCode]
      properties:
        topicId:
          type: string
          format: uuid
        draftType:
          $ref: '#/components/schemas/DraftType'
        tone:
          type: string
        languageCode:
          type: string
    UpdateDraftRequest:
      type: object
      additionalProperties: false
      properties:
        title:
          type: string
        outline:
          type: array
          items:
            type: object
            additionalProperties: true
        body:
          type: string
        facts:
          type: array
          items:
            type: object
            additionalProperties: true
        sourceRefs:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/SourceReference'
        tone:
          type: string
    RenderDraftRequest:
      type: object
      additionalProperties: false
      required: [template, pageCount, size]
      properties:
        template:
          type: string
        pageCount:
          type: integer
          minimum: 1
          maximum: 20
        size:
          type: string
          pattern: '^[0-9]{3,5}x[0-9]{3,5}$'
...
    ChannelVariant:
      type: object
      additionalProperties: false
      required:
        [id, draftId, platform, hashtags, assetBundle, publishMode, status, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        draftId:
          type: string
          format: uuid
        platform:
          $ref: '#/components/schemas/PlatformType'
        title:
          type: string
        caption:
          type: string
        hashtags:
          type: array
          items:
            type: string
        coverAssetId:
          type: string
          format: uuid
        assetBundle:
          type: array
          items:
            type: string
            format: uuid
        publishMode:
          type: string
          enum: [semi_auto, direct_api]
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      description: Platform-specific publish payload. The generic draft body is intentionally not exposed here.
    CreateChannelVariantsRequest:
      type: object
      additionalProperties: false
      required: [platforms]
      properties:
        platforms:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/PlatformType'
    UpdateChannelVariantRequest:
      type: object
      additionalProperties: false
      properties:
        title:
          type: string
        caption:
          type: string
        hashtags:
          type: array
          items:
            type: string
        coverAssetId:
          type: string
          format: uuid
        assetBundle:
          type: array
          items:
            type: string
            format: uuid
        status:
          type: string
    ReviewTask:
      type: object
      additionalProperties: false
      required: [id, draftId, status, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        draftId:
          type: string
          format: uuid
        reviewerId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/ReviewStatus'
        comments:
          type: string
        scheduledAt:
          type: string
          format: date-time
        approvedAt:
          type: string
          format: date-time
        rejectedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateReviewTaskRequest:
      type: object
      additionalProperties: false
      required: [draftId]
      properties:
        draftId:
          type: string
          format: uuid
        reviewerId:
          type: string
          format: uuid
    ApproveReviewTaskRequest:
      type: object
      additionalProperties: false
      required: [selectedPlatforms]
      properties:
        scheduledAt:
          type: string
          format: date-time
        selectedPlatforms:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/PlatformType'
    RejectReviewTaskRequest:
      type: object
      additionalProperties: false
      required: [comments]
      properties:
        comments:
          type: string
          minLength: 1
    PublishTask:
      type: object
      additionalProperties: false
      required:
        [id, channelVariantId, platform, status, retryCount, responsePayload, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        channelVariantId:
          type: string
          format: uuid
        platform:
          $ref: '#/components/schemas/PlatformType'
        status:
          $ref: '#/components/schemas/PublishStatus'
        scheduledAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        publishedAt:
          type: string
          format: date-time
        remotePostId:
          type: string
        retryCount:
          type: integer
          minimum: 0
        responsePayload:
          type: object
          additionalProperties: true
        errorMessage:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreatePublishTaskRequest:
      type: object
      additionalProperties: false
      required: [channelVariantId, platform]
      properties:
        channelVariantId:
          type: string
          format: uuid
        platform:
          $ref: '#/components/schemas/PlatformType'
        scheduledAt:
          type: string
          format: date-time
...
    DashboardOverview:
      type: object
      additionalProperties: false
      required:
        [rawItemsCount, topicsCount, draftsGenerated, pendingReviews, scheduledPublishes, publishedToday]
      properties:
        rawItemsCount:
          type: integer
          minimum: 0
        topicsCount:
          type: integer
          minimum: 0
        draftsGenerated:
          type: integer
          minimum: 0
        pendingReviews:
          type: integer
          minimum: 0
        scheduledPublishes:
          type: integer
          minimum: 0
        publishedToday:
          type: integer
          minimum: 0
    JobAcceptedResponse:
      type: object
      additionalProperties: false
      required: [jobId, status]
      properties:
        jobId:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued]
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    DraftListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Draft'
...
    ChannelVariantListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ChannelVariant'
    ReviewTaskListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ReviewTask'
    PublishTaskListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PublishTask'
...